+++ examples/multi-agent-systems/room_mas/view_simulation_in_dashboard.py
@@ -1,8 +1,10 @@
 """View the room_mas simulation in a dashboard.
 
-To test the dashboard functionality, make sure you run the room_mas long enough (either using real-time simulation, or a very long simulation duration).
-Also, you need the optional dependency 'interactive'.
+To test the dashboard functionality, make sure you run the room_mas long enough
+(either using real-time simulation, or a very long simulation duration). Also,
+you need the optional dependency 'interactive'.
 
+Make sure the simulation is already running and saving to a csv file!
 
 
 """
@@ -14,4 +16,4 @@ from agentlib.utils import simulator_dashboard
 
 if __name__ == "__main__":
     os.chdir(Path(__file__).parent)
-    simulator_dashboard("res_room1.csv")
+    simulator_dashboard("res_room1.csv")  # accepts multiple files as *args